+++ packages/tailwindcss-language-server/tests/completions/completions.test.js
@@ -310,8 +310,8 @@ withFixture('v4/basic', (c) => {
     let result = await completion({ lang, text, position, settings })
     let textEdit = expect.objectContaining({ range: { start: position, end: position } })
 
-    expect(result.items.length).toBe(12376)
-    expect(result.items.filter((item) => item.label.endsWith(':')).length).toBe(220)
+    expect(result.items.length).toBe(12398)
+    expect(result.items.filter((item) => item.label.endsWith(':')).length).toBe(224)
     expect(result).toEqual({
       isIncomplete: false,
       items: expect.arrayContaining([
@@ -475,6 +475,36 @@ withFixture('v4/basic', (c) => {
 
     expect(result.items.length).toBe(23)
     expect(result.items.filter((item) => item.label.startsWith('--')).length).toBe(23)
+  })
+
+  test.concurrent('@slot is suggeted inside @variant', async ({ expect }) => {
+    let result = await completion({
+      lang: 'css',
+      text: '@',
+      position: { line: 0, character: 1 },
+    })
+
+    // Make sure `@slot` is NOT suggested by default
+    expect(result.items.length).toBe(10)
+    expect(result.items).not.toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({ kind: 14, label: '@slot', sortText: '-0000000' }),
+      ]),
+    )
+
+    result = await completion({
+      lang: 'css',
+      text: '@variant foo {\n@',
+      position: { line: 1, character: 1 },
+    })
+
+    // Make sure `@slot` is suggested
+    expect(result.items.length).toBe(11)
+    expect(result.items).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({ kind: 14, label: '@slot', sortText: '-0000000' }),
+      ]),
+    )
   })
 
   test.concurrent('resolve', async ({ expect }) => {