+++ packages/tailwindcss-language-server/tests/fixtures/v4/css-loading-js/cjs/my-plugin.cjs
@@ -1,0 +1,16 @@
+const plugin = require('tailwindcss/plugin')
+
+module.exports = plugin(
+  () => {
+    //
+  },
+  {
+    theme: {
+      extend: {
+        colors: {
+          'cjs-from-plugin': 'black',
+        },
+      },
+    },
+  },
+)